Refactor all SmartRice to PalayLab package
