+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 26
     defaultConfig {
-        applicationId "io.github.projectbukirin.smartrice"
+        applicationId "io.github.projectbukirin.palaylab"
         minSdkVersion 23
         targetSdkVersion 26
         versionCode 1
