+++ app/src/androidTest/java/io/github/projectbukirin/palaylab/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package io.github.projectbukirin.smartrice;
+package io.github.projectbukirin.palaylab;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("io.github.projectbukirin.smartrice", appContext.getPackageName());
+        assertEquals("io.github.projectbukirin.palaylab", appContext.getPackageName());
     }
 }
