+++ app/src/main/java/io/github/projectbukirin/palaylab/About.java
@@ -1,4 +1,4 @@
-package io.github.projectbukirin.smartrice;
+package io.github.projectbukirin.palaylab;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
