+++ app/src/main/java/io/github/projectbukirin/palaylab/PestsDefinition.java
@@ -1,4 +1,4 @@
-package io.github.projectbukirin.smartrice;
+package io.github.projectbukirin.palaylab;
 
 import android.graphics.Color;
 import android.graphics.Typeface;
