+++ app/src/main/java/io/github/projectbukirin/palaylab/Statistics.java
@@ -1,4 +1,4 @@
-package io.github.projectbukirin.smartrice;
+package io.github.projectbukirin.palaylab;
 
 import android.content.DialogInterface;
 import android.content.Intent;
