+++ app/src/main/java/io/github/projectbukirin/palaylab/TensorFlowImageClassifier.java
@@ -1,4 +1,4 @@
-package io.github.projectbukirin.smartrice;
+package io.github.projectbukirin.palaylab;
 
 import android.annotation.SuppressLint;
 import android.content.res.AssetFileDescriptor;
