+++ app/src/test/java/io/github/projectbukirin/palaylab/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package io.github.projectbukirin.smartrice;
+package io.github.projectbukirin.palaylab;
 
 import org.junit.Test;
